hvc_sysfs: change hyp trace log size to uint32_t
The type of the hypevisor log size is changing to uint32_t.
Modifying the hypervisor control driver accordingly.

JIRA VFND-2811

Reviewed-on: http://git-master/r/1253012
Reviewed-on: http://git-master/r/1314846
GVS: Gerrit_Virtual_Submit
Tested-by: Aniruddha Banerjee <aniruddhab@nvidia.com>

+++ include/soc/tegra/virt/syscalls.h
@@ -159,7 +159,7 @@ struct hyp_server_page {
 
 	/* hypervisor trace log */
 	uint64_t log_ipa;
-	uint64_t log_size;
+	uint32_t log_size;
 };
 
 /* For backwards compatibility, alias the old name for hyp_server_name. */
